adding a more difficult function. Refactoring and cleaning code. Adding results info.

+++ Individual.java
@@ -2,7 +2,7 @@ public class Individual {
     private int phenotype;
     private int bitFieldSize;
     private int[] genotype;
-    private int fitness;
+    private double fitness;
     private double chanceToSurvive;
 
     public Individual(int bitFieldSize) {
@@ -28,11 +28,11 @@ public class Individual {
         return this.genotype;
     }
 
-    public void setFitness(int fitness) {
+    public void setFitness(double fitness) {
         this.fitness = fitness;
     }
 
-    public int getFitness() {
+    public double getFitness() {
         return this.fitness;
     }
